+++ tests/test_fields.py
@@ -3,9 +3,12 @@ import enum
 
 from unittest import TestCase
 
+from django.core.exceptions import ImproperlyConfigured
+from rest_framework import fields
 from rest_framework.exceptions import ValidationError
+from rest_framework.serializers import Serializer
 
-from rest_framework_dataclasses.fields import DefaultDecimalField, EnumField, IterableField, MappingField
+from rest_framework_dataclasses.fields import DefaultDecimalField, EnumField, IterableField, MappingField, UnionField
 
 
 class FieldTest(TestCase):
@@ -60,3 +63,51 @@ class FieldTest(TestCase):
         ordered_field = MappingField(container=collections.OrderedDict)
         ordered_values = {'foo': 'bar', 'abc': 'def'}
         self.assertEqual(ordered_field.to_internal_value(ordered_values), collections.OrderedDict(ordered_values))
+
+    def test_union_field(self):
+        class A:
+            def __init__(self, a):
+                self.a = a
+
+        class B:
+            def __init__(self, b):
+                self.b = b
+
+        class ASerializer(Serializer):
+            a = fields.CharField()
+
+        class BSerializer(Serializer):
+            b = fields.IntegerField()
+
+        # Note that A/B are regular classes and serializers, so internal value is a dict and not a dataclass.
+        ab_field = UnionField({A: ASerializer, B: BSerializer}, nest_value=False)
+        self.assertEqual(ab_field.to_internal_value({'type': 'A', 'a': 'a'}), {'a': 'a'})
+        self.assertEqual(ab_field.to_internal_value({'type': 'B', 'b': 42}), {'b': 42})
+        self.assertEqual(ab_field.to_representation(A('a')), {'type': 'A', 'a': 'a'})
+        self.assertEqual(ab_field.to_representation(B(42)), {'type': 'B', 'b': 42})
+
+        strint_field = UnionField({str: fields.CharField, int: fields.IntegerField}, nest_value=True)
+        self.assertEqual(strint_field.to_internal_value({'type': 'str', 'value': 'hello'}), 'hello')
+        self.assertEqual(strint_field.to_internal_value({'type': 'int', 'value': 42}), 42)
+        self.assertEqual(strint_field.to_representation('hello'), {'type': 'str', 'value': 'hello'})
+        self.assertEqual(strint_field.to_representation(42), {'type': 'int', 'value': 42})
+
+        with self.assertRaises(ValidationError):
+            strint_field.to_internal_value({'value': 3})
+        with self.assertRaises(ValidationError):
+            strint_field.to_internal_value({'type': 'int'})
+        with self.assertRaises(ValidationError):
+            strint_field.to_internal_value({'type': 'int', 'value': 3.5})
+        with self.assertRaises(ValidationError):
+            strint_field.to_internal_value({'type': 'float', 'value': 3})
+
+        invalid_field = UnionField({str: fields.CharField, int: fields.IntegerField}, nest_value=False)
+        with self.assertRaises(ImproperlyConfigured):
+            invalid_field.to_representation(42)
+
+        renamed_field = UnionField({int: fields.IntegerField, float: fields.FloatField},
+                                   nest_value=True,
+                                   discriminator_field_name='renamed_type',
+                                   value_field_name='renamed_value')
+        self.assertEqual(renamed_field.to_internal_value({'renamed_type': 'float', 'renamed_value': 1.2}), 1.2)
+        self.assertEqual(renamed_field.to_representation(3.4), {'renamed_type': 'float', 'renamed_value': 3.4})